+++ script/BilibiliCover/BilibiliCover.user.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name            B站封面获取
-// @version         5.9.0.20230106
+// @version         5.10.0.20230311
 // @namespace       laster2800
 // @author          Laster2800
 // @description     获取B站各播放页及直播间封面，支持手动及实时预览等多种模式，支持点击下载、封面预览、快速复制，可高度自定义
@@ -19,7 +19,7 @@
 // @require         https://greasyfork.org/scripts/431998-userscriptapidom/code/UserscriptAPIDom.js?version=1005139
 // @require         https://greasyfork.org/scripts/432000-userscriptapimessage/code/UserscriptAPIMessage.js?version=1095149
 // @require         https://greasyfork.org/scripts/432002-userscriptapiwait/code/UserscriptAPIWait.js?version=1129540
-// @require         https://greasyfork.org/scripts/432003-userscriptapiweb/code/UserscriptAPIWeb.js?version=1095148
+// @require         https://greasyfork.org/scripts/432003-userscriptapiweb/code/UserscriptAPIWeb.js?version=1160007
 // @grant           GM_download
 // @grant           GM_notification
 // @grant           GM_xmlhttpRequest
@@ -31,9 +31,9 @@
 // @grant           GM_unregisterMenuCommand
 // @connect         api.bilibili.com
 // @run-at          document-start
-// @compatible      edge 版本不小于 85
-// @compatible      chrome 版本不小于 85
-// @compatible      firefox 版本不小于 90
+// @compatible      edge 版本不小于 93
+// @compatible      chrome 版本不小于 93
+// @compatible      firefox 版本不小于 92
 // ==/UserScript==
 
 (function() {
@@ -51,7 +51,7 @@
       display: block;
       width: 100%;
     }
-  `.trim().replace(/\s+/g, ' ')
+  `.trim().replaceAll(/\s+/g, ' ')
 
   const gm = {
     id: gmId,
@@ -394,7 +394,7 @@
           if (result === '') {
             result = gm.configMap.customModeStyle.default
           } else {
-            result = result.replace(/\s+/g, ' ')
+            result = result.replaceAll(/\s+/g, ' ')
           }
           gm.config.customModeStyle = result
           GM_setValue('customModeStyle', result)